feat: separate search from logic from <SearchInput/> view component
